new.sh: read the file name from the first line
read the file name from the first line instead of prompting the user for
it. this makes the script simpler and less interactive, it also gives
the user the possibility to change the note's name at any point of the
editing process.

Hm, might be a good idea to extend this to the rest of the utils, the
first line is always the file name